refactor etherscan aip to enable verifying through the api
