+++ apps/etherscan/src/app/views/VerifyView.tsx
@@ -5,11 +5,9 @@ import {
 } from "@remixproject/plugin"
 import { Formik, ErrorMessage, Field } from "formik"
 
-import { getNetworkName, getEtherScanApi, getReceiptStatus } from "../utils"
 import { SubmitButton } from "../components"
 import { Receipt } from "../types"
-import { CompilationResult } from "@remixproject/plugin-api"
-import axios from 'axios'
+import { verify } from "../utils/verify"
 
 interface Props {
   client: PluginClient
@@ -24,43 +22,7 @@ interface FormValues {
   contractAddress: string
 }
 
-export const getContractFileName = (
-  compilationResult: CompilationResult,
-  contractName: string
-) => {
-  const compiledContracts = compilationResult.contracts
-  let fileName = ""
 
-  for (const file of Object.keys(compiledContracts)) {
-    for (const contract of Object.keys(compiledContracts[file])) {
-      if (contract === contractName) {
-        fileName = file
-        break
-      }
-    }
-  }
-  return fileName
-}
-
-export const getContractMetadata = (
-  compilationResult: CompilationResult,
-  contractName: string
-) => {
-  const compiledContracts = compilationResult.contracts
-  let contractMetadata = ""
-
-  for (const file of Object.keys(compiledContracts)) {
-    for (const contract of Object.keys(compiledContracts[file])) {
-      if (contract === contractName) {
-        contractMetadata = compiledContracts[file][contract].metadata
-        if (contractMetadata) {
-          break
-        }
-      }
-    }
-  }
-  return contractMetadata
-}
 
 export const VerifyView: React.FC<Props> = ({
   apiKey,
@@ -80,117 +42,20 @@ export const VerifyView: React.FC<Props> = ({
       throw new Error("no compilation result available")
     }
 
-    const contractArguments = values.contractArguments.replace("0x", "")
-
-    const verify = async (
-      apiKeyParam: string,
-      contractAddress: string,
-      contractArgumentsParam: string,
-      contractName: string,
-      compilationResultParam: any
-    ) => {
-      const network = await getNetworkName(client)
-      if (network === "vm") {
-        return "Cannot verify in the selected network"
-      }
-      const etherscanApi = getEtherScanApi(network)
-
-      try {
-        const contractMetadata = getContractMetadata(
-          compilationResultParam.data,
-          contractName
-        )
-
-        if (!contractMetadata) {
-          return "Please recompile contract"
-        }
-        
-        const contractMetadataParsed = JSON.parse(contractMetadata)
-
-        const fileName = getContractFileName(
-          compilationResultParam.data,
-          contractName
-        )
-
-        const jsonInput = {
-          language: 'Solidity',
-          sources: compilationResultParam.source.sources,
-          settings: {
-            optimizer: {
-              enabled: contractMetadataParsed.settings.optimizer.enabled,
-              runs: contractMetadataParsed.settings.optimizer.runs
-            }
-          }
-        }
-
-        const data: { [key: string]: string | any } = {
-          apikey: apiKeyParam, // A valid API-Key is required
-          module: "contract", // Do not change
-          action: "verifysourcecode", // Do not change
-          codeformat: "solidity-standard-json-input",
-          contractaddress: contractAddress, // Contract Address starts with 0x...
-          sourceCode: JSON.stringify(jsonInput),
-          contractname: fileName + ':' + contractName,
-          compilerversion: `v${contractMetadataParsed.compiler.version}`, // see http://etherscan.io/solcversions for list of support versions
-          constructorArguements: contractArgumentsParam, // if applicable
-        }
-
-        const body = new FormData()
-        Object.keys(data).forEach((key) => body.append(key, data[key]))
-
-        client.emit("statusChanged", {
-          key: "loading",
-          type: "info",
-          title: "Verifying ...",
-        })
-        const response = await axios.post(etherscanApi, body)
-        const { message, result, status } = await response.data
-
-        if (message === "OK" && status === "1") {
-          resetAfter10Seconds()
-          const receiptStatus = await getReceiptStatus(
-            result,
-            apiKey,
-            etherscanApi
-          )
-
-          onVerifiedContract({
-            guid: result,
-            status: receiptStatus,
-          })
-          return `Contract verified correctly <br> Receipt GUID ${result}`
-        }
-        if (message === "NOTOK") {
-          client.emit("statusChanged", {
-            key: "failed",
-            type: "error",
-            title: result,
-          })
-          resetAfter10Seconds()
-        }
-        return result
-      } catch (error) {
-        console.error(error)
-        setResults("Something wrong happened, try again")
-      }
-    }
-
-    const resetAfter10Seconds = () => {
-      setTimeout(() => {
-        client.emit("statusChanged", { key: "none" })
-        setResults("")
-      }, 10000)
-    }
+    const contractArguments = values.contractArguments.replace("0x", "")    
 
     const verificationResult = await verify(
       apiKey,
       values.contractAddress,
       contractArguments,
       values.contractName,
-      compilationResult
+      compilationResult,
+      client,
+      onVerifiedContract,
+      setResults,
     )
 
-    setResults(verificationResult)
+    setResults(verificationResult.message)
   }
 
   return (
